+++ detector/src/main/java/nl/littlerobots/rxlint/RxJava2SubscriberCheck.java
@@ -6,17 +6,68 @@ import com.intellij.psi.PsiParameter;
 import com.intellij.psi.PsiType;
 import com.intellij.psi.util.TypeConversionUtil;
 
+import org.jetbrains.uast.UCallExpression;
+import org.jetbrains.uast.UExpression;
+import org.jetbrains.uast.UQualifiedReferenceExpression;
+import org.jetbrains.uast.USimpleNameReferenceExpression;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 public class RxJava2SubscriberCheck implements SubscribeDetector.SubscriberCheck {
+
+    private static final String OBSERVABLE_TYPE = "io.reactivex.Observable";
+    private static final String FLOWABLE_TYPE = "io.reactivex.Flowable";
+    private static final String COMPLETABLE_TYPE = "io.reactivex.Completable";
+    private static final String SINGLE_TYPE = "io.reactivex.Single";
+    private static final String MAYBE_TYPE = "io.reactivex.Maybe";
+
+    private static Map<String, List<String>> ERROR_HANDLING_OPERATORS = new HashMap<String, List<String>>(5);
+
+    static {
+        ERROR_HANDLING_OPERATORS.put(OBSERVABLE_TYPE, Collections.singletonList("onErrorReturnItem("));
+        ERROR_HANDLING_OPERATORS.put(FLOWABLE_TYPE, Collections.singletonList("onErrorReturnItem("));
+        ERROR_HANDLING_OPERATORS.put(COMPLETABLE_TYPE, Collections.singletonList("onErrorComplete()"));
+        ERROR_HANDLING_OPERATORS.put(MAYBE_TYPE, Arrays.asList("onErrorComplete()", "onErrorReturnItem("));
+    }
+
     @Override
-    public boolean isMissingOnError(PsiMethod method) {
+    public boolean isMissingOnError(UCallExpression node, PsiMethod method) {
         PsiClass clz = method.getContainingClass();
         String type = clz.getQualifiedName();
-        if ("io.reactivex.Observable".equals(type) ||
-                "io.reactivex.Flowable".equals(type) ||
-                "io.reactivex.Completable".equals(type) ||
-                "io.reactivex.Single".equals(type) ||
-                "io.reactivex.Maybe".equals(type)) {
-            return hasNoErrorHandling(method);
+        if (OBSERVABLE_TYPE.equals(type) ||
+                FLOWABLE_TYPE.equals(type) ||
+                COMPLETABLE_TYPE.equals(type) ||
+                SINGLE_TYPE.equals(type) ||
+                MAYBE_TYPE.equals(type)) {
+            return hasNoErrorHandling(method) && canProduceError(node, type);
+        }
+        return false;
+    }
+
+    private boolean canProduceError(UCallExpression node, String type) {
+        return !isErrorSuppressingOperator(node.getReceiver(), type);
+    }
+
+    private boolean isErrorSuppressingOperator(UExpression receiver, String type) {
+        List<String> methods = ERROR_HANDLING_OPERATORS.get(type);
+        if (methods == null || methods.isEmpty()) {
+            return false;
+        }
+        if (receiver == null) {
+            return false;
+        }
+        if (receiver instanceof UQualifiedReferenceExpression && !(receiver instanceof USimpleNameReferenceExpression)) {
+            //FIXME Change to a proper way to get the method name and params
+            String selector = ((UQualifiedReferenceExpression) receiver).getSelector().toString();
+            for (String method : methods) {
+                if (selector.startsWith(method)) {
+                    return true;
+                }
+            }
         }
         return false;
     }